+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 0.4 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Do not let themes and categories end up in the catalog index when
+  the behavior is not activated for a portal_type.
+  [maurits]
 
 
 0.3 (2014-02-24)